workflowLlamaIndex/requirements.txt
langgraph
langchain_openai
tavily-python
langchain_community
langgraph-cli
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
asyncer==0.0.7
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
certifi==2024.8.30
chainlit==1.2.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
docopt==0.6.2
executing==2.1.0
fastapi==0.112.4
fastjsonschema==2.20.0
filetype==1.2.0
frozenlist==1.4.1
googleapis-common-protos
grpcio==1.66.1
grpcio-tools
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.4.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.1
langchain-openai==0.2.0
langchain-qdrant==0.1.4
langchain-text-splitters==0.3.0
langsmith==0.1.121
Lazify==0.4.0
literalai==0.0.607
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.46.0
opentelemetry-api>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-exporter-otlp-proto-common>=1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
opentelemetry-exporter-otlp-proto-http>=1.27.0
opentelemetry-instrumentation>=0.48b0
opentelemetry-proto
opentelemetry-sdk>=1.27.0
opentelemetry-semantic-conventions>=0.48b0
orjson==3.10.7
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.47
protobuf>=3.20.0,<6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.11.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.37.2
syncer==2.0.3
tenacity==8.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.26.0
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.20.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.11.1
zipp==3.20.2
faiss-cpu
llama-index
llama-index-tools-tavily-research
llama-index-utils-workflow

#%pip install -qU llama-index llama-index-tools-tavily-research llama-index-utils-workflow